Separate authorization utils and server functions
